requirements.txt
torch>=2.0.0
pytorch-lightning>=2.0.0
transformers>=4.30.0
wandb>=0.15.0
hydra-core>=1.3.0
einops>=0.6.0
torchmetrics>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
seaborn>=0.12.0
matplotlib>=3.5.0
